+++ python/cuml/ensemble/randomforestregressor.pyx
@@ -298,6 +298,66 @@ class RandomForestRegressor(BaseRandomForestModel, RegressorMixin):
         self.treelite_serialized_model = None
         self.n_cols = None
 
+    def n_nodes_(self):
+        """
+        Calculates the number of nodes in the trained RF model
+
+        Returns
+        ----------
+        num_nodes : int, number of nodes
+        """
+        cdef RandomForestMetaData[float, int] *rf_forest = \
+            <RandomForestMetaData[float, int]*><uintptr_t> self.rf_forest
+        cdef RandomForestMetaData[double, int] *rf_forest64 = \
+            <RandomForestMetaData[double, int]*><uintptr_t> self.rf_forest64
+
+        if self.dtype == np.float32:
+            num_nodes = calc_num_nodes(rf_forest)
+        else:
+            num_nodes = calc_num_nodes(rf_forest64)
+
+        return num_nodes
+
+    def forest_info_(self, forest_param="threshold"):
+        """
+        Returns the requested forest parameter values as a 2d vector. The
+        2 dimensions of the vector represent (num_trees, num_splits).
+
+        Parameters
+        ----------
+        forest_param : string (default=threshold)
+            This parameter accepts the following values:
+
+            "threshold":  Returns a 2d vector of vaules used as threshold at
+            each node.
+            "best_metric": Returns a 2d vector of the the best metric value
+            for each node.
+            "column_id": Returns a 2d vector of column/feature id used to
+            create the split at each node.
+
+        Returns
+        ----------
+        Forest_info 2d vector containing requested values of the forest
+        """
+        select_feature = {"threshold": 0,
+                          "best_metric": 1,
+                          "column_id": 2}
+        cdef RandomForestMetaData[float, float] *rf_forest = \
+            <RandomForestMetaData[float, float]*><uintptr_t> self.rf_forest
+        cdef RandomForestMetaData[double, double] *rf_forest64 = \
+            <RandomForestMetaData[double, double]*><uintptr_t> self.rf_forest64
+
+        if self.dtype == np.float32:
+            forest_info_vector = obtain_forest_info(
+                rf_forest,
+                <int> select_feature[forest_param])
+        else:
+            forest_info_vector = obtain_forest_info(
+                rf_forest64,
+                <int> select_feature[forest_param])
+
+        return forest_info_vector
+
     def convert_to_treelite_model(self):
         """
         Converts the cuML RF model to a Treelite model